+++ tests/Faithlife.Utility.Tests/EnumerableUtilityTests.cs
@@ -6,6 +6,9 @@ using System.Linq;
 using MyLib;
 using NUnit.Framework;
 
+// Don't warn for testing obsolete methods
+#pragma warning disable CS0618
+
 namespace Faithlife.Utility.Tests
 {
 	[TestFixture]
@@ -195,31 +198,31 @@ namespace Faithlife.Utility.Tests
 		[Test]
 		public void DistinctBy()
 		{
-			Assert.Throws<ArgumentNullException>(() => ((IEnumerable<int>) null!).DistinctBy(i => i));
-			Assert.Throws<ArgumentNullException>(() => new[] { 1 }.DistinctBy((Func<int, int>) null!));
+			Assert.Throws<ArgumentNullException>(() => EnumerableUtility.DistinctBy((IEnumerable<int>) null!, i => i));
+			Assert.Throws<ArgumentNullException>(() => EnumerableUtility.DistinctBy(new[] { 1 }, (Func<int, int>) null!));
 
-			CollectionAssert.AreEqual(new[] { 2.1, 1, 2, 3, 1.4 }.DistinctBy(d => Math.Floor(d)), new[] { 2.1, 1, 3 });
-			CollectionAssert.AreEqual(new double[] { 1, 2, 1, 3, 1 }.DistinctBy(d => Math.Floor(d)), new double[] { 1, 2, 3 });
-			CollectionAssert.AreEqual(new double[] { 1, 2, 1, 3, 1 }.DistinctBy(d => Math.Floor(d), null), new double[] { 1, 2, 3 });
+			CollectionAssert.AreEqual(EnumerableUtility.DistinctBy(new[] { 2.1, 1, 2, 3, 1.4 }, d => Math.Floor(d)), new[] { 2.1, 1, 3 });
+			CollectionAssert.AreEqual(EnumerableUtility.DistinctBy(new double[] { 1, 2, 1, 3, 1 }, d => Math.Floor(d)), new double[] { 1, 2, 3 });
+			CollectionAssert.AreEqual(EnumerableUtility.DistinctBy(new double[] { 1, 2, 1, 3, 1 }, d => Math.Floor(d), null), new double[] { 1, 2, 3 });
 
-			CollectionAssert.AreEqual(new[] { "a", "ab", "abc", "abcd" }.DistinctBy(s => s.Length), new[] { "a", "ab", "abc", "abcd" });
-			CollectionAssert.AreEqual(new[] { "a", "b", "c", "ab", "abc", "abcd", "bcd" }.DistinctBy(s => s.Length), new[] { "a", "ab", "abc", "abcd" });
+			CollectionAssert.AreEqual(EnumerableUtility.DistinctBy(new[] { "a", "ab", "abc", "abcd" }, s => s.Length), new[] { "a", "ab", "abc", "abcd" });
+			CollectionAssert.AreEqual(EnumerableUtility.DistinctBy(new[] { "a", "b", "c", "ab", "abc", "abcd", "bcd" }, s => s.Length), new[] { "a", "ab", "abc", "abcd" });
 		}
 
 		[Test]
 		public void DistinctByComparerArgumentNull()
 		{
-			Assert.Throws<ArgumentNullException>(() => ((IEnumerable<int>) null!).DistinctBy(i => i, EqualityComparer<int>.Default));
-			Assert.Throws<ArgumentNullException>(() => new[] { 1 }.DistinctBy(null!, EqualityComparer<int>.Default));
+			Assert.Throws<ArgumentNullException>(() => EnumerableUtility.DistinctBy((IEnumerable<int>) null!, i => i, EqualityComparer<int>.Default));
+			Assert.Throws<ArgumentNullException>(() => EnumerableUtility.DistinctBy(new[] { 1 }, null!, EqualityComparer<int>.Default));
 		}
 
 		[Test]
 		public void DistinctByComparer()
 		{
 			CollectionAssert.AreEqual(new[] { "a", "ab" },
-				new[] { "a", "ab", "abc", "abcd" }.DistinctBy(s => s.Length, new OddEvenEqualityComparer()));
+				EnumerableUtility.DistinctBy(new[] { "a", "ab", "abc", "abcd" }, s => s.Length, new OddEvenEqualityComparer()));
 			CollectionAssert.AreEqual(new[] { "a", "ab" },
-				new[] { "a", "b", "c", "ab", "abc", "abcd", "bcd" }.DistinctBy(s => s.Length, new OddEvenEqualityComparer()));
+				EnumerableUtility.DistinctBy(new[] { "a", "b", "c", "ab", "abc", "abcd", "bcd" }, s => s.Length, new OddEvenEqualityComparer()));
 		}
 
 		private class OddEvenEqualityComparer : EqualityComparer<int>
@@ -550,10 +553,8 @@ namespace Faithlife.Utility.Tests
 		[Test]
 		public void TakeLastArguments()
 		{
-#pragma warning disable CS0618 // Type or member is obsolete
 			Assert.Throws<ArgumentNullException>(() => EnumerableUtility.TakeLast((IEnumerable<int>) null!, 1));
 			Assert.Throws<ArgumentOutOfRangeException>(() => EnumerableUtility.TakeLast(new[] { 1 }, -1));
-#pragma warning restore CS0618 // Type or member is obsolete
 
 			// As .NET Standard 2.0
 			Assert.Throws<ArgumentNullException>(() => MyClass.DoTakeLast((IEnumerable<int>) null!, 1));
@@ -563,9 +564,7 @@ namespace Faithlife.Utility.Tests
 		[Test]
 		public void TakeLastZero()
 		{
-#pragma warning disable CS0618 // Type or member is obsolete
 			CollectionAssert.AreEqual(new int[0], EnumerableUtility.TakeLast(new[] { 1 }, 0));
-#pragma warning restore CS0618 // Type or member is obsolete
 
 			// As .NET Standard 2.0
 			CollectionAssert.AreEqual(new int[0], MyClass.DoTakeLast(new[] { 1 }, 0));
@@ -574,9 +573,7 @@ namespace Faithlife.Utility.Tests
 		[Test]
 		public void TakeLastEmpty()
 		{
-#pragma warning disable CS0618 // Type or member is obsolete
 			CollectionAssert.AreEqual(new int[0], EnumerableUtility.TakeLast(Enumerable.Empty<int>(), 100));
-#pragma warning restore CS0618 // Type or member is obsolete
 
 			// As .NET Standard 2.0
 			CollectionAssert.AreEqual(new int[0], MyClass.DoTakeLast(Enumerable.Empty<int>(), 100));
@@ -586,9 +583,7 @@ namespace Faithlife.Utility.Tests
 		public void TakeLastPartialSequence()
 		{
 			var sequence = new EnumerableMonitor<int>(Enumerable.Range(1, 100));
-#pragma warning disable CS0618 // Type or member is obsolete
 			CollectionAssert.AreEqual(new[] { 98, 99, 100 }, EnumerableUtility.TakeLast(sequence, 3));
-#pragma warning restore CS0618 // Type or member is obsolete
 			Assert.AreEqual(100, sequence.RequestCount);
 
 			// As .NET Standard 2.0
@@ -600,9 +595,7 @@ namespace Faithlife.Utility.Tests
 		public void TakeLastEntireSequence()
 		{
 			var sequence = new EnumerableMonitor<int>(Enumerable.Range(1, 3));
-#pragma warning disable CS0618 // Type or member is obsolete
 			CollectionAssert.AreEqual(new[] { 1, 2, 3 }, EnumerableUtility.TakeLast(sequence, 5));
-#pragma warning restore CS0618 // Type or member is obsolete
 			Assert.AreEqual(3, sequence.RequestCount);
 
 			// As .NET Standard 2.0